KNOX-2811 - Renamed the main angular token generation component
