+++ knox-token-generation-ui/package.json
@@ -6,7 +6,7 @@
     "start": "ng serve --base-href /gateway/homepage/token-generation/ --deploy-url /gateway/homepage/token-generation/ --verbose=true",
     "build": "ng build",
     "build-prod": "ng build --prod",
-    "lint": "tslint \"token-gen/**/*.ts\""
+    "lint": "tslint \"token-generation/**/*.ts\""
   },
   "private": true,
   "dependencies": {
