+++ knox-token-generation-ui/token-generation/app/token-generation.component.html
