+++ knox-token-generation-ui/token-generation/index.html
@@ -27,6 +27,6 @@
 	</script>
 </head>
 <body class="login">
-    <app-token-gen></app-token-gen>
+    <app-token-generation></app-token-generation>
 </body>
 </html>
